# app/static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "description": "GT20 Team Final Project of Backend Track B-)",
        "version": "0.0.1",
        "title": "Startup Campus Backend Final Project",
        "contact": {
            "name": "Team GT20",
            "email": "gt20@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://fashion-campus-gt20.my.id:5000"
        }
    ],
    "components": {
        "securitySchemes": {
            "Fashion-Campus-Auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header",
                "name": "JWT-FC"
            }
        },
...
    },
    "paths": {
        "/image/{urlPath}": {
            "get": {
                "tags": [
                    "UNIVERSAL"
                ],
                "summary": "Get Image to Show In The Platform",
                "description": "Get image from google cloud storage",
                "parameters": [
                    {
                        "name": "urlPath",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "image/png": {
                                "schema": {
                                    "type": "image/png",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "displaying image": {
                                        "description": "Responses for failed get an image from Google Cloud Storage",
                                        "value": {
                                            "message": "Error, image not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "PRODUCT LIST"
                ],
                "summary": " Get an available product from website's catalog",
                "parameters": [
                    {
                        "name": "page",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page_size",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "condition",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_name",
                        "description": "aaa",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Get Product List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string"
                                                    },
                                                    "title": {
                                                        "type": "string"
                                                    },
                                                    "price": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        },
                                        "total_rows": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "examples": {
                                    "Get Product List": {
                                        "description": "Example of successfully get product list",
                                        "value": {
                                            "data": [
                                                {
                                                    "id": "11987fc5-c3b9-427a-9999-9dcca19cc649",
                                                    "images": "/static/barcelona_2011.png",
                                                    "title": "Barcelona Home Jersey Champions League Winner",
                                                    "price": 1500000
                                                }
                                            ],
                                            "total_rows": 4
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "Get Product List": {
                                        "description": "Example of failed getting product list",
                                        "value": {
                                            "message": "Error, product not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ADMIN PAGE"
                ],
                "summary": "Add product by Admin",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "array"
                                    },
                                    "condition": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                }
                            },
                            "examples": {
                                "Request Body": {
                                    "description": "Example of correct request body",
                                    "value": {
                                        "product_name": "Barcelona 2011/2012 Long Sleeve Original Jersey",
                                        "description": "Jersey where Barcelona win Champions League and Club World Cup Trophy",
                                        "images": [
                                            "Barcelona_Home_2011/2012.png",
                                            "Barcelona_Away_2011/2012.png"
                                        ],
                                        "condition": "New",
                                        "category": "Shirt",
                                        "price": 1500000
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Responses for successfully adding product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Example of success add product from admin",
                                        "value": {
                                            "message": "Product added"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            },
            "put": {
                "tags": [
                    "ADMIN PAGE"
                ],
                "summary": "Update product by Admin",
                "security": [
                    {
                        "Fashion-Campus-Auth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "array"
                                    },
                                    "condition": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                }
                            },
                            "examples": {
                                "Request Body": {
                                    "description": "Example of correct request body",
                                    "value": {
                                        "product_name": "Real Madrid Long Sleeve Original Jersey",
                                        "description": "Jersey where Real Madird win Champions League in 2018",
                                        "images": [
                                            "Real_Madrid_Home_2018/2019.png",
                                            "Real_Madrid_Away_2018/2019.png"
                                        ],
                                        "condition": "New",
                                        "category": "Shirt",
                                        "price": 1500000
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Responses for successfully updating product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "success": {
                                        "description": "Example of success updating product from admin",
                                        "value": {
                                            "message": "Product updated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/UnauthorizedError"
                    }
                }
            }
        },
...
        "/products/search_image": {
            "post": {
                "tags": [
                    "PRODUCT LIST"
                ],
                "summary": "Search a product by image using base64 encoded string",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/vnd.github.base64": {
                            "schema": {
                                "type": "string",
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "required": true
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "aaa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "PRODUCT DETAIL PAGE"
                ],
                "summary": "Get detail information from an item",
                "parameters": [
                    {
                        "name": "id",
                        "description": "aaa",
                        "required": false,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Get Product Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "size": {
                                            "type": "string"
                                        },
                                        "product_detail": {
                                            "type": "string"
                                        },
                                        "price": {
                                            "type": "integer"
                                        },
                                        "images_url": {
                                            "type": "string"
                                        },
                                        "category_id": {
                                            "type": "string"
                                        },
                                        "category_name": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}